+++ persistent/tests/test_persistence.py
@@ -1697,13 +1697,16 @@ class _Persistent_Base(object):
         self.assertEqual(candidate._p_state, GHOST)
         self.assertEqual(candidate.set_by_new, 1)
 
+    # The number 12345678 as a p64, 8-byte string
+    _PACKED_OID = b'\x00\x00\x00\x00\x00\xbcaN'
+    # The number 12345678 printed in hex
+    _HEX_OID = '0xbc614e'
+
     def _normalize_repr(self, r):
         # addresses
-        r = re.sub(r'0x[0-9a-fA-F]*', '0xdeadbeef', r)
+        r = re.sub(r'at 0x[0-9a-fA-F]*', 'at 0xdeadbeef', r)
         # Python 3.7 removed the trailing , in exception reprs
         r = r.replace("',)", "')")
-        # Python 2 doesn't have a leading b prefix for byte literals
-        r = r.replace("oid '", "oid b'")
         return r
 
     def _normalized_repr(self, o):
@@ -1730,12 +1733,12 @@ class _Persistent_Base(object):
 
     def test_repr_oid_no_jar(self):
         p = self._makeOne()
-        p._p_oid = b'12345678'
+        p._p_oid = self._PACKED_OID
 
         result = self._normalized_repr(p)
         self.assertEqual(
             result,
-            "<persistent.Persistent object at 0xdeadbeef oid b'12345678'>")
+            "<persistent.Persistent object at 0xdeadbeef oid " + self._HEX_OID + ">")
 
     def test_repr_no_oid_repr_jar_raises_exception(self):
         p = self._makeOne()
@@ -1758,7 +1761,17 @@ class _Persistent_Base(object):
         class BadOID(bytes):
             def __repr__(self):
                 raise Exception("oid repr failed")
-        p._p_oid = BadOID(b'12345678')
+
+        # Our OID is bytes, 8 bytes long. We don't call its repr.
+        p._p_oid = BadOID(self._PACKED_OID)
+
+        result = self._normalized_repr(p)
+        self.assertEqual(
+            result,
+            "<persistent.Persistent object at 0xdeadbeef oid " + self._HEX_OID + ">")
+
+        # Anything other than 8 bytes, though, we do.
+        p._p_oid = BadOID(b'1234567')
 
         result = self._normalized_repr(p)
         self.assertEqual(
@@ -1772,7 +1785,7 @@ class _Persistent_Base(object):
         class BadOID(bytes):
             def __repr__(self):
                 raise Exception("oid repr failed")
-        p._p_oid = BadOID(b'12345678')
+        p._p_oid = BadOID(b'1234567')
 
         class Jar(object):
             def __repr__(self):
@@ -1806,12 +1819,17 @@ class _Persistent_Base(object):
                 raise BaseException("oid repr failed")
         p._p_oid = BadOID(b'12345678')
 
+        # An 8 byte byte string doesn't have repr called.
+        repr(p)
+
+        # Anything other does.
+        p._p_oid = BadOID(b'1234567')
         with self.assertRaisesRegex(BaseException, 'oid repr failed'):
             repr(p)
 
     def test_repr_oid_and_jar(self):
         p = self._makeOne()
-        p._p_oid = b'12345678'
+        p._p_oid = self._PACKED_OID
 
         class Jar(object):
             def __repr__(self):
@@ -1822,7 +1840,7 @@ class _Persistent_Base(object):
         result = self._normalized_repr(p)
         self.assertEqual(
             result,
-            "<persistent.Persistent object at 0xdeadbeef oid b'12345678' in <SomeJar>>")
+            "<persistent.Persistent object at 0xdeadbeef oid " + self._HEX_OID + " in <SomeJar>>")
 
     def test__p_repr(self):
         class P(self._getTargetClass()):
@@ -1842,12 +1860,12 @@ class _Persistent_Base(object):
             "<persistent.tests.test_persistence.P object at 0xdeadbeef"
             " _p_repr Exception('_p_repr failed')>")
 
-        p._p_oid = b'12345678'
+        p._p_oid = self._PACKED_OID
         result = self._normalized_repr(p)
         self.assertEqual(
             result,
-            "<persistent.tests.test_persistence.P object at 0xdeadbeef oid b'12345678'"
-            " _p_repr Exception('_p_repr failed')>")
+            "<persistent.tests.test_persistence.P object at 0xdeadbeef oid " + self._HEX_OID
+            + " _p_repr Exception('_p_repr failed')>")
 
         class Jar(object):
             def __repr__(self):
@@ -1857,8 +1875,8 @@ class _Persistent_Base(object):
         result = self._normalized_repr(p)
         self.assertEqual(
             result,
-            "<persistent.tests.test_persistence.P object at 0xdeadbeef oid b'12345678'"
-            " in <SomeJar> _p_repr Exception('_p_repr failed')>")
+            "<persistent.tests.test_persistence.P object at 0xdeadbeef oid " + self._HEX_OID
+            + " in <SomeJar> _p_repr Exception('_p_repr failed')>")
 
     def test__p_repr_in_instance_ignored(self):
         class P(self._getTargetClass()):